Refactor 3D scene management into dedicated module
